requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
loguru>=0.7.0
Pillow>=10.0.0
mss>=9.0.1  # For fast screenshot capture
streamlit>=1.30.0  # Management Dashboard
discord.py>=2.3.0  # Discord Bot Integration
pandas>=2.0.0  # For data display in dashboard
goose-ai>=0.9.0  # Coding assistant integration
duckduckgo-search>=6.0.0  # Web search functionality


